Implemented edit and delete on project view
